requirements.txt
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
google-api-core==1.20.1
google-auth==1.17.2
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.29.0
idna==2.9
MouseInfo==0.1.2
numpy==1.18.5
opencv-contrib-python==4.2.0.34
Pillow==7.1.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.48
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.3
PyTweening==1.0.3
pytz==2020.1
pywin32==227
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.9
